Split SampleVocabulary.tsx into per-component files
The 469-line file held two fetch components and six presentational ones.
WordRow (+WordTagBadges), TableSkeleton, PaginatedVocabulary (+shortcut
hint), and TextbookVocabulary each get their own file; SampleVocabulary.tsx
keeps only its namesake. Pure relocation — logic moved to
lib/sample-vocabulary in an earlier commit.
