+++ src/components/sections/KanjiDetails/Details.tsx
@@ -7,7 +7,8 @@ import { FrequencyInfo } from "./FrequencyInfo";
 import { General } from "./General";
 import { KanjiKeyboardShortcuts } from "./KanjiKeyboardShortcuts";
 import { ReadingFrequencyCategory } from "./ReadingFrequencyCategory";
-import { SampleVocabulary, TextbookVocabulary } from "./SampleVocabulary";
+import { SampleVocabulary } from "./SampleVocabulary";
+import { TextbookVocabulary } from "./TextbookVocabulary";
 import { Badge } from "@/components/ui/badge";
 import { outLinks } from "@/lib/external-links";
 import { ExternalKanjiLinks } from "@/components/common/ExternalKanjiLinks";
